requirements.txt
pillow
requests
tqdm
torch
torchvision
opencv-python
matplotlib
pyyaml
psutil
scipy
pandas